+++ apps/newsletters-ui/src/app/App.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 import { Api } from './Api';
-import { ButtonContainer } from './ButtonContainer';
+import { ButtonContainer } from './Components/ButtonContainer';
 import { DataClientTest } from './DataClientTest';
 
 const StyledApp = styled.div`
